+++ crates/services/producer/src/block_producer.rs
@@ -21,17 +21,17 @@ use fuel_core_types::{
         primitives::DaBlockHeight,
     },
     fuel_asm::Word,
-    fuel_tx::{
-        Receipt,
-        Transaction,
-    },
+    fuel_tx::Transaction,
     fuel_types::{
         BlockHeight,
         Bytes32,
     },
     services::{
         block_producer::Components,
-        executor::UncommittedResult,
+        executor::{
+            TransactionExecutionStatus,
+            UncommittedResult,
+        },
     },
     tai64::Tai64,
 };
@@ -178,15 +178,15 @@ where
     Executor: ports::DryRunner + 'static,
 {
     // TODO: Support custom `block_time` for `dry_run`.
-    /// Simulate a transaction without altering any state. Does not aquire the production lock
+    /// Simulates multiple transactions without altering any state. Does not acquire the production lock.
     /// since it is basically a "read only" operation and shouldn't get in the way of normal
     /// production.
     pub async fn dry_run(
         &self,
-        transaction: Transaction,
+        transactions: Vec<Transaction>,
         height: Option<BlockHeight>,
         utxo_validation: Option<bool>,
-    ) -> anyhow::Result<Vec<Receipt>> {
+    ) -> anyhow::Result<Vec<TransactionExecutionStatus>> {
         let height = height.unwrap_or_else(|| {
             self.view_provider
                 .latest_height()
@@ -194,7 +194,6 @@ where
                 .expect("It is impossible to overflow the current block height")
         });
 
-        let is_script = transaction.is_script();
         // The dry run execution should use the state of the blockchain based on the
         // last available block, not on the upcoming one. It means that we need to
         // use the same configuration as the last block -> the same DA height.
@@ -203,25 +202,31 @@ where
         let header = self._new_header(height, Tai64::now())?;
         let component = Components {
             header_to_produce: header,
-            transactions_source: transaction,
+            transactions_source: transactions.clone(),
             gas_limit: u64::MAX,
         };
 
         let executor = self.executor.clone();
+
         // use the blocking threadpool for dry_run to avoid clogging up the main async runtime
-        let res: Vec<_> =
-            tokio_rayon::spawn_fifo(move || -> anyhow::Result<Vec<Receipt>> {
-                Ok(executor
-                    .dry_run(component, utxo_validation)?
-                    .into_iter()
-                    .flatten()
-                    .collect())
+        let tx_statuses = tokio_rayon::spawn_fifo(
+            move || -> anyhow::Result<Vec<TransactionExecutionStatus>> {
+                Ok(executor.dry_run(component, utxo_validation)?)
+            },
+        )
+        .await?;
+
+        if transactions
+            .iter()
+            .zip(tx_statuses.iter())
+            .any(|(transaction, tx_status)| {
+                transaction.is_script() && tx_status.receipts.is_empty()
             })
-            .await?;
-        if is_script && res.is_empty() {
-            return Err(anyhow!("Expected at least one set of receipts"))
+        {
+            Err(anyhow!("Expected at least one set of receipts"))
+        } else {
+            Ok(tx_statuses)
         }
-        Ok(res)
     }
 }